move node data
