+++ Assets/MistNet/Runtime/Scripts/ConnectionSelector/NodeData.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 914592d2e5c40a749be043c0c72517b7
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
